Item price calculation

+++ src/main/java/com/example/g2gcalculator/dto/ItemResponse.java
@@ -6,10 +6,10 @@ import lombok.Builder;
 public record ItemResponse(
         Integer auctionHouseId,
         Integer itemId,
-        Integer minBuyout,
+        Long minBuyout,
         Integer quantity,
-        Integer marketValue,
-        Integer historical,
+        Long marketValue,
+        Long historical,
         Integer numAuctions
 ) {
 }
